feat: user joins room test with an invalid room
