+++ src/types/errors.ts
@@ -1,0 +1,5 @@
+const errorsMessages = {
+  INVALID_ROOM: 'invalid room provided'
+};
+
+export default errorsMessages;
