+++ src/types/index.ts
@@ -1,1 +1,2 @@
-export * from './server';
+export * from './serverTypes';
+export * from './errors';
